requirements.txt
Django==5.0.4
djongo==1.3.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pymongo==4.7.2
dnspython==2.6.1
pytz==2024.1